feat(apps/web): Refactoring
